Removed repeated code for email confirmation
