+++ voluncheer/profiles/urls.py
@@ -3,7 +3,7 @@ from django.urls import path
 from profiles.views.home import home
 from profiles.views.profile import ProfileView
 from profiles.views.profile import profile_update
-from profiles.views.volunteers import activate
+from profiles.views.activate_email import activate
 
 urlpatterns = [
     path("", home, name="home"),
