+++ voluncheer/templates/registration/template_activate_account.html
@@ -1,4 +1,4 @@
-{% extends 'base.html' %} 
+
 {% load crispy_forms_tags %} 
 {% block content %}
